requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.30,<1.0

# Typing and settings
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
email-validator>=2.0,<3.0

# Database and migrations
SQLAlchemy>=2.0,<3.0
greenlet>=3.0,<4.0
asyncpg>=0.29,<1.0
alembic>=1.13,<2.0

# HTTP client for tests and integrations
httpx>=0.27,<1.0
pytest>=8.2,<9.0
pytest-asyncio>=0.23,<1.0

# Logging and observability
structlog>=24.1,<25.0
sentry-sdk[fastapi]>=2.5,<3.0

# Performance and env
orjson>=3.10,<4.0
python-dotenv>=1.0,<2.0
PyJWT>=2.8,<3.0
boto3>=1.35,<2.0

# Tasks and queues
redis>=5.0,<6.0
rq>=1.16,<2.0

# Dev tools
ruff==0.6.4
mypy==1.11.2
pre-commit==3.8.0


